Fixed serious CPU spikes in certain conditions
Fixed serious CPU spikes in certain conditions
1.Now if house items > LIMIT in depot of player and player will cast
leavehouse the items will be LOST it is to prevent LONG saves of players
that are having thousands items in depot...
2.You can not send parcel if receiver depot is full.
3.Fixed protocolgame issue that was abusable.

+++ src/protocolgame.cpp
@@ -1016,8 +1016,8 @@ void ProtocolGame::parseAutoWalk(NetworkMessage& msg)
 {
 	// first we get all directions...
 	std::list<Direction> path;
-	size_t numdirs = msg.GetByte();
-	for (size_t i = 0; i < numdirs; ++i) {
+	uint8_t numdirs = msg.GetByte();
+	for (uint8_t i = 0; i < numdirs; ++i) {
 		uint8_t rawdir = msg.GetByte();
 		Direction dir = SOUTH;
